requirements.txt

Mezzanine==3.1.6
Pillow==2.5.1
django_debug_toolbar==1.2.1
psycopg2==2.5.3
python_memcached
django_extensions==1.3.8
django_ses==0.6.0
django-simple-captcha==0.4.2
boto==2.31.1
easy-thumbnails==1.4


